requirements.txt
fastapi
uvicorn
gitpython
anthropic
python-dotenv

